requirements.txt
fastapi
uvicorn
python-dotenv
openai==2.9.0
httpx
